Updated all components to use the style component Put components in their own directory Put component styles in their own file
